requirements.txt
annotated-types==0.7.0
attrs==24.3.0
Automat==24.8.1
cachetools==5.5.0
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.0
constantly==23.10.4
contourpy==1.3.1
cryptography==44.0.0
cssselect==1.2.0
cycler==0.12.1
defusedxml==0.7.1
Faker==33.1.0
filelock==3.16.1
fonttools==4.55.3
google-ai-generativelanguage==0.6.10
google-api-core==2.24.0
google-api-python-client==2.155.0
google-auth==2.37.0
google-auth-httplib2==0.2.0
google-generativeai==0.8.3
googleapis-common-protos==1.66.0
greenlet==3.1.1
grpcio==1.68.1
grpcio-status==1.68.1
httplib2==0.22.0
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
itemadapter==0.10.0
itemloaders==1.3.2
jmespath==1.0.1
kiwisolver==1.4.8
lxml==5.3.0
matplotlib==3.10.0
numpy==2.2.0
packaging==24.2
pandas==2.2.3
parsel==1.9.1
pillow==11.0.0
Protego==0.3.1
proto-plus==1.25.0
protobuf==5.29.1
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pydantic==2.10.3
pydantic_core==2.27.1
PyDispatcher==2.0.7
pyOpenSSL==24.3.0
pyparsing==3.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
queuelib==1.7.0
requests==2.32.3
requests-file==2.1.0
rsa==4.9
Scrapy==2.12.0
service-identity==24.2.0
setuptools==75.6.0
six==1.17.0
SQLAlchemy==2.0.36
tldextract==5.1.3
tqdm==4.67.1
Twisted==24.11.0
typing_extensions==4.12.2
tzdata==2024.2
uritemplate==4.1.1
urllib3==2.2.3
w3lib==2.2.1
zope.interface==7.2
